Move statik filesystem implemention to subpackage of statik package.
