+++ handler.go
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:generate statik -src=./webui
-
 package pilosa
 
 import (
