+++ handler_test.go
@@ -31,7 +31,7 @@ import (
 	"github.com/pilosa/pilosa"
 	"github.com/pilosa/pilosa/internal"
 	"github.com/pilosa/pilosa/pql"
-	"github.com/pilosa/pilosa/statikfs"
+	statik "github.com/pilosa/pilosa/statik/filesystem"
 	"github.com/pilosa/pilosa/test"
 )
 
@@ -1854,7 +1854,7 @@ func TestHandler_WebUI(t *testing.T) {
 	h := test.NewHandler()
 	h.Holder = hldr.Holder
 	h.Cluster = test.NewCluster(1)
-	h.FileSystem = &statikfs.FileSystem{}
+	h.FileSystem = &statik.FileSystem{}
 
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("GET", "/", nil))
