+++ server/server.go
@@ -35,7 +35,7 @@ import (
 	"github.com/pilosa/pilosa"
 	"github.com/pilosa/pilosa/gcnotify"
 	"github.com/pilosa/pilosa/gossip"
-	"github.com/pilosa/pilosa/statikfs"
+	statik "github.com/pilosa/pilosa/statik/filesystem"
 	"github.com/pilosa/pilosa/statsd"
 )
 
@@ -193,7 +193,7 @@ func (m *Command) SetupServer() error {
 	m.Server.Cluster.RemoteClient = c
 
 	// Statik file system.
-	m.Server.Handler.FileSystem = &statikfs.FileSystem{}
+	m.Server.Handler.FileSystem = &statik.FileSystem{}
 
 	// Default coordintor to port 0 when not specified so that coordinator
 	// can be set to the value of server.URI after server binds to a port.
