+++ statik/.gitignore
@@ -1,1 +1,1 @@
-statik.go
+/statik.go
