+++ statik/doc.go
@@ -1,3 +1,4 @@
 // Package statik contains static assets for the Web UI. `go generate` will
 // produce statik.go, which is ignored by git.
+//go:generate statik -src=../webui -dest=..
 package statik
